FIX: The retract error generated by the travel
When the retraction setting is larger than the retraction during the material change, the travel triggered during the material change will make up for the retraction interpolation, but the retracted extruder head will be the wrong one. By moving the travel part in the extruder switch Gcode to after the traditional head change (where m_curr_extruder_id in writer() is updated), the old ID is avoided during travel.

jira: STUDIO-14764
Change-Id: I8af8f66af180f01de2fef5760601ee54e73548c4